Corrected argument lists for calling FXMessageBox class methods (such as FXMessageBox.error).
The problem was that GCC now expects the "format" string argument
to be a string literal and it can't be a variable. So I modified the
way we call those so that we pass the string literal "%s" for that
argument, and then pass the user-supplied string argument as the following
argument.