Add cross-compiled binaries for Ruby 1.8 and 1.9 to the gem spec.
By default it looks like rake-compiler doesn't modify the gem spec
appropriately when we're cross-compiling for multiple Ruby versions.
Added some lines to remove "lib/fox16.so" from the gem spec's list
of files to include, and then added "lib/1.8/fox16.so" and "lib/1.9/fox16.so"
back to the list, so that we get the correct result.