Fix Segfault while GC'ing FXWindows
FXWindow destructor calls recalc() and changeFocus() of it's parent windows.
Since these methods are routed back to Ruby code, but calling Ruby code from
GC isn't a good idea, we mark the parent window as "in_gc", so that it will
ignore recalc() and changeFocus() calls completely.

The parent window should also be scheduled to be free'd. In the other case,
the child window would have been marked as used.