Downgrade zlib version for Windows binary gem to 1.2.7.3.
This is necessary for RubyInstaller versions based on MSYS1 (Ruby < 2.4).

RubyInstaller comes bundled with zlib-1.2.7. Since the zlib1.dll is stored in the
same directory as ruby.exe, it is preferred in the Windows DLL load order.
The zlib1.dll that is bundled with fxruby is not loaded at runtime.

If we build fxruby with a newer zlib version, libpng will make use of the new
function inflateValidate() introduced in zlib-1.2.8, which results in a DLL
load error at runtime. We avoid this by sticking at zlib-1.2.7.