Add a 3rd type of FXRbObjects which is used for callbacks.
Since commit 54a1170 we handle arguments to blocks (callbacks) as borrowed objects and decouple
libfox objects from the corresponding FXRuby wrapper object after the call, if it's a borrowed object.
The conflict arises when the objects is already a borrowed object and is additionally passed as block argument.
In this case the object is deregistered at the end of the callback and accidently gets free'd by the GC later on.

With block arguments as dedicated object type we can differentiate between objects wrapped for one callback only and objects that are for longer use.

Fixes #39